first commit 9

+++ app/controllers/posts_controller.rb
@@ -1,5 +1,6 @@
 class PostsController < ApplicationController
 	protect_from_forgery :only => ["create"]
+	#protect_from_forgery except: :create
 	
   def index
 	  @posts = Post.all
@@ -9,7 +10,10 @@ class PostsController < ApplicationController
   end
 
   def create
-	  post=Post.create(title: params[:title], content: params[:content])
+	if user_signed_in?
+		writer = current_user.email
+	end
+	  post=Post.create(title: params[:title], content: params[:content], writer: writer)
 	  redirect_to "/posts/show/#{post.id}"
   end
 
@@ -29,7 +33,14 @@ class PostsController < ApplicationController
 
   def destroy
 	  post = Post.find(params[:id])
-	  post.destroy
-	  redirect_to "/posts/index"
+	  
+	  if user_signed_in?
+	  	if post.writer == current_user.email
+		  post.destroy
+		  redirect_to "/posts/index"
+	  	else
+		end
+		  
+	  end
   end
 end
